Rename the file
A: dWorldEditor, not dRegions.
B: Putting things in a WIP folder breaks git. Files should stay in one
folder all their lifetime, with one name. This is so git commits are
properly tracked to the file, rather than only showing one commit 'move
file to release folder' and 0 of the actual commits that went into
working it.